tilføjet hash transfer method, fixet fejl i file transfer method og message transfer method. mere info i plugin entries. fixet fejl med at beans ikke blev lavet ordentlig

+++ src/main/java/me/alek/serversecurity/socket/methods/SocketFileTransferMethod.java
@@ -1,21 +1,26 @@
 package me.alek.serversecurity.socket.methods;
 
-import me.alek.serversecurity.bot.SingletonBotInitializer;
-import me.alek.serversecurity.socket.ISocketTransferMethod;
-import me.alek.serversecurity.socket.SocketContext;
+import me.alek.serversecurity.bot.DiscordBot;
+import me.alek.serversecurity.socket.INestableSocketTransferMethod;
+import me.alek.serversecurity.socket.SocketPipelineContext;
 
 import java.io.DataInputStream;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.net.ServerSocket;
 
-public class SocketFileTransferMethod implements ISocketTransferMethod {
+public class SocketFileTransferMethod implements INestableSocketTransferMethod<File> {
 
+    private File file = null;
 
     @Override
-    public void handle(ServerSocket serverSocket, InputStream stream, SocketContext context) {
+    public File handle(ServerSocket serverSocket, InputStream stream, SocketPipelineContext context) {
         try {
             String fileName = context.getStorage().poll(String.class);
+
+            file = new File("tmp/" + fileName);
+            if (file.exists()) return file;
 
             DataInputStream dataStream = new DataInputStream(stream);
             FileOutputStream fileOutputStream = new FileOutputStream("tmp/" + fileName);
@@ -31,12 +36,20 @@ public class SocketFileTransferMethod implements ISocketTransferMethod {
             fileOutputStream.close();
             dataStream.close();
 
-            SingletonBotInitializer.log("Succesfully transfered file " + fileName + " with " + totalBytes + " total bytes");
+            DiscordBot.log("Successfully transfered file " + fileName + " with " + totalBytes + " total bytes");
         }
         catch (Exception ex) {
 
-            ex.printStackTrace();
-            SingletonBotInitializer.log("Error occurred in file transfer: " + ex.getMessage());
+            if (!context.hasClosed()) {
+                ex.printStackTrace();
+                DiscordBot.log("Error occurred in file transfer: " + ex.getMessage());
+            }
         }
+        return file;
+    }
+
+    @Override
+    public void putIntoSharedContext(SocketPipelineContext context) {
+        context.getStorage().put(File.class, file);
     }
 }
